sidecar/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
websockets>=12.0
instructor>=1.15.1
pydantic>=2.7.0
polars>=1.40.0
lancedb>=0.30.2
sentence-transformers>=2.5.1
fastexcel>=0.10.0
einops
pytest
passlib[bcrypt]
python-jose[cryptography]
python-multipart
